Create README.md
When using a LinkedList, Insertion Sort is much more efficient, and this is easy to see from the graph provided above. A LinkedList allows for elements, or Nodes, to be placed in the middle much more easily, as the memory addresses for values have no relation to one another. Therefore, pointers can easily be redirected to change the order of a list. When in a vector or array, doing so is a lot more laborious on the computer, as each individual element must be pushed back one index in the list. The difference between the two become a lot more apparent when lists become thousands of Nodes long, and thousands of arrays are required to be pushed back.